+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 (TBD)
 
 * Add CallableMetadata.returnType
+* Add ClassMetadata.isData
 * Update to Kotlin 1.2.0
 
 ## 1.0.0 (Feb 15, 2016)